# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/login": {
            "post": {
                "description": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "要注册的用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "description": "前端发送的 Json 对象",
                        "name": "registerParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterParam"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/user/sendCode": {
            "post": {
                "description": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "parameters": [
                    {
                        "description": "前端发送的 Json 对象",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IdentifyingCode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/user/testAuth": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "访问该接口的时候, 用户必须是已经登录成功",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "访问该接口的时候, 用户必须是已经登录成功",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.IdentifyingCode": {
            "type": "object",
            "properties": {
                "identity": {
                    "description": "身份标志;1:雇主，2:开发者,3:忘记密码",
                    "type": "string"
                },
                "phone": {
                    "description": "电话",
                    "type": "string"
                }
            }
        },
        "dto.RegisterParam": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "identity_tag": {
                    "description": "身份标志;1:雇主，2:开发者",
                    "type": "integer"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "电话",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "swagger.Http400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "swagger.Http404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Page Not Found"
                }
            }
        },
        "swagger.Http500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "服务器内部错误"
                }
            }
        },
        "swagger.HttpOk": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "接口调用成功"
                }
            }
        }
    }
}